Separated general security service methods into its own service
