+++ ripple-security/src/main/java/org/rippleosi/login/service/LoginService.java
@@ -86,38 +86,4 @@ public class LoginService {
             manager.save(true, profile);
         }
     }
-
-    public ResponseEntity<String> generateRedirectResponseEntity(final String defaultUrl, final Map<String, String> params,
-                                                                 final HttpStatus httpStatus) {
-        CommonHelper.assertNotNull("defaultUrl", defaultUrl);
-        CommonHelper.assertNotBlank("defaultUrl", defaultUrl);
-
-        final String redirectUrl = getExpandedUrl(defaultUrl, params);
-        final HttpHeaders httpHeaders = new HttpHeaders();
-
-        try {
-            final URI redirectPage = new URI(redirectUrl);
-            httpHeaders.setLocation(redirectPage);
-        }
-        catch (final URISyntaxException e) {
-            LOGGER.warn("The security service has failed to redirect to the requested URL.");
-            LOGGER.debug("The security service has failed to redirect to the requested URL.", e);
-        }
-
-        return new ResponseEntity<>(httpHeaders, httpStatus);
-    }
-
-    private String getExpandedUrl(final String defaultUrl, final Map<String, String> params) {
-        if (params == null) {
-            return defaultUrl;
-        }
-
-        final UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(defaultUrl);
-
-        for (final Map.Entry<String, String> entry : params.entrySet()) {
-            builder.queryParam(entry.getKey(), entry.getValue());
-        }
-
-        return builder.toUriString();
-    }
 }
